[Chore] rename to `testService.ts`
